% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcL1AV.R
\name{calcL1AV}
\alias{calcL1AV}
\title{Calculate L1 Acceptance Value}
\usage{
calcL1AV(x, T = 100)
}
\arguments{
\item{x}{A vector containing the individual assay values of 10 dosage units}

\item{T}{The target assay value.  T=100 in most circumstances.}
}
\value{
List containing the L1 Acceptance Value and a logical
indicating whether the sample passed.
}
\description{
Calculate the L1 acceptance value from the assay values of a 10 dosage unit
sample.
}
\examples{
sample <- seq(95, 104)
calcL1AV(sample)
}
