% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcL2AV.R
\name{calcL2AV}
\alias{calcL2AV}
\title{Calculate L2 Acceptance Value}
\usage{
calcL2AV(sample10, sample20, T = 100)
}
\arguments{
\item{sample10}{A vector containing the individual assay values of 10 dosage
units.}

\item{sample20}{A vector containing the individual assay values of an
additional 20 dosage units.}

\item{T}{The target assay value.  T=100 in most circumstances.}
}
\value{
List the L2 Acceptance Value and a logical
indicating whether the sample passed.
}
\description{
Calculate the L2 acceptance value from the assay values of a 30-unit sample
of dosage units.  Usually applied after an L1 failure.
}
\examples{
sample1 <- seq(95, 104)
sample2 <- seq(95, 104.5, by=0.5)
calcL2AV(sample1, sample2)
}
