% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uspCUTest.R
\name{uspCUTest}
\alias{uspCUTest}
\title{USP Content Uniformity test}
\usage{
uspCUTest(sample10, sample20 = NULL, T = 100)
}
\arguments{
\item{sample10}{A vector containing the individual assay values of 10 dosage
units.}

\item{sample20}{A vector containing the individual assay values of an
additional 20 dosage units.}

\item{T}{The target assay value.  T=100 in most circumstances.}
}
\value{
List the L1 Acceptance Value, a logical indicating
whether the sample passed L1, the L2 Acceptance Value and a
logical indicating whether the sample passed
}
\description{
USP Content Uniformity test
}
